+++ src/lib.rs
@@ -90,7 +90,7 @@ pub use network::{
     NetworkClient, NetworkConfig, NetworkEvent, NetworkSystem, RemoteEntities, SnapshotBuffer,
 };
 pub use particle::{Particle, ParticleBurst, ParticleEmitter, ParticleSystem};
-pub use pathfinding::{find_path, PathGrid};
+pub use pathfinding::{find_path, find_path_diagonal, PathGrid};
 #[cfg(not(target_arch = "wasm32"))]
 pub use physics::{
     BodyHandle, CharacterController, ColliderHandle, CollisionEvent, CollisionGroups, JointHandle,